ports/psoc6: Moved all module files to root dir.
