+++ ports/psoc6/machine_adc.c
