+++ ports/psoc6/machine_adc.h
