+++ ports/psoc6/machine_adcblock.c
