+++ ports/psoc6/machine_spi.c
