+++ ports/psoc6/machine_timer.c
